gcc: be finer-grained and less brute-forcing in doc installation.
The installation used to have a list of files that was the set of all
possible files to install, no matter the target packages. This resulted
in very few hits in the list each time, something hard to read and
debug. Be finer-grained and give the exact list of files to install for
each gccpackage.