libmpc: patch -no-undefined in the m4/am files and run autoreconf -fiv.
Previous patches directly changed configure and src/Makefile.in. I think
that was because I had a mistake in a regexp and I couldn't manage to
patch otherwise; that or I hadn't seen it was acinclude.m4 I had to
change.

Patch removes -no-undefined from acinclude.m4 (which gets into
configure) and instead puts it in libmpc_la_LDFLAGS in src/Makefile.am

After that it's possible to run autoreconf -fiv to regenerate other
parts of the build system so that libtool correctly handles our -L flag;
the one in the tarball fails to find gmp and mpfr.