libtheora: build cross-compiled an executable it then tried to run.
John L. Poole on the mailing-list had an issue with rebuilding libtheora.
Turns out it was failing in several places but went unnoticed because of
the lack of "set -e".

John's issue was caused by an old automake on his system and the lack of
"set -e" made the build fail only quite long after the actual issue. Adding
it uncovered other issues.

"make" failed but for a rule about doc/examples and "make install" went
far enough to provide the files needed for build and running.

Fixing this requires the ax_proc_cc_for_build.m4 file which provides the
machinery needed to detect (but not build) a compiler suitable for the
build machine and we also need to patch doc/spec/Makefile.am.